Get things working
Finally it is working /o/
Even though the code is a bit messy, the app is now working with the creation of tree, watering and shaman (which currently will update the tree) and also the logout button.

There is room for a ton of improvements, but now we can focus on that!